+++ src/MenuListInspector.js
@@ -8,17 +8,19 @@ import type MenuEvent from './events/MenuEvent';
 import ChosenEvent from './events/ChosenEvent';
 import type {Direction, Rect} from './types';
 
-export type MenuListInspectorContext = {
+export type MenuListInspectorContextValue = {
   registerMenuList(menuList: MenuList): void;
   unregisterMenuList(menuList: MenuList): void;
   dispatchEvent(event: MenuEvent): void;
 };
 
+export const MenuListInspectorContext = React.createContext<?MenuListInspectorContextValue>(null);
+
 export type Props = {
-  onItemChosen?: ?(event: ChosenEvent) => void;
-  onLeftPushed?: ?(event: MenuEvent) => void;
-  onRightPushed?: ?(event: MenuEvent) => void;
-  children?: ReactNode;
+  onItemChosen?: (event: ChosenEvent) => void;
+  onLeftPushed?: (event: MenuEvent) => void;
+  onRightPushed?: (event: MenuEvent) => void;
+  children: ReactNode;
 };
 
 export default class MenuListInspector extends React.Component<Props> {
@@ -30,52 +32,43 @@ export default class MenuListInspector extends React.Component<Props> {
     children: PropTypes.element
   };
 
-  static childContextTypes = {
-    menuListInspector: PropTypes.object
-  };
-
-  static contextTypes = {
-    menuListInspector: PropTypes.object
-  };
+  static contextType = MenuListInspectorContext;
 
   _descendantMenuLists: Array<MenuList> = [];
 
-  _parentCtx(): ?MenuListInspectorContext {
-    return this.context.menuListInspector;
+  _parentCtx(): ?MenuListInspectorContextValue {
+    return this.context;
   }
 
-  getChildContext(): Object {
-    const menuListInspector: MenuListInspectorContext = {
-      registerMenuList: (menuList: MenuList) => {
-        this._descendantMenuLists.push(menuList);
-      },
-      unregisterMenuList: (menuList: MenuList) => {
-        const i = this._descendantMenuLists.indexOf(menuList);
-        if (i < 0) throw new Error('MenuList not registered');
-        this._descendantMenuLists.splice(i, 1);
-      },
-      dispatchEvent: (event: MenuEvent) => {
-        switch (event.type) {
-        case 'chosen':
-          /*:: if (!(event instanceof ChosenEvent)) throw new Error(); */
-          if (this.props.onItemChosen) this.props.onItemChosen(event);
-          break;
-        case 'left':
-          if (this.props.onLeftPushed) this.props.onLeftPushed(event);
-          break;
-        case 'right':
-          if (this.props.onRightPushed) this.props.onRightPushed(event);
-          break;
-        }
-        if (event.cancelBubble) return;
-        const parentCtx = this._parentCtx();
-        if (parentCtx) {
-          parentCtx.dispatchEvent(event);
-        }
+  _menuListInspectorContext: MenuListInspectorContextValue = {
+    registerMenuList: (menuList: MenuList) => {
+      this._descendantMenuLists.push(menuList);
+    },
+    unregisterMenuList: (menuList: MenuList) => {
+      const i = this._descendantMenuLists.indexOf(menuList);
+      if (i < 0) throw new Error('MenuList not registered');
+      this._descendantMenuLists.splice(i, 1);
+    },
+    dispatchEvent: (event: MenuEvent) => {
+      switch (event.type) {
+      case 'chosen':
+        /*:: if (!(event instanceof ChosenEvent)) throw new Error(); */
+        if (this.props.onItemChosen) this.props.onItemChosen(event);
+        break;
+      case 'left':
+        if (this.props.onLeftPushed) this.props.onLeftPushed(event);
+        break;
+      case 'right':
+        if (this.props.onRightPushed) this.props.onRightPushed(event);
+        break;
       }
-    };
-    return {menuListInspector};
-  }
+      if (event.cancelBubble) return;
+      const parentCtx = this._parentCtx();
+      if (parentCtx) {
+        parentCtx.dispatchEvent(event);
+      }
+    }
+  };
 
   moveCursor(direction: Direction, prevCursorLocation: ?Rect): boolean {
     const menuList = this._descendantMenuLists[0];
@@ -96,6 +89,10 @@ export default class MenuListInspector extends React.Component<Props> {
   }
 
   render() {
-    return this.props.children;
+    return (
+      <MenuListInspectorContext.Provider value={this._menuListInspectorContext}>
+        {this.props.children}
+      </MenuListInspectorContext.Provider>
+    );
   }
 }