+++ src/SubMenuItem.js
@@ -5,13 +5,14 @@ import kefirBus from 'kefir-bus';
 import type {Bus} from 'kefir-bus';
 import kefirStopper from 'kefir-stopper';
 import React from 'react';
-import type {Node as ReactNode, Element as ReactElement} from 'react';
+import type {Node as ReactNode} from 'react';
 import PropTypes from 'prop-types';
 import pointRectDistance from './lib/pointRectDistance';
 
 import MenuListInspector from './MenuListInspector';
 import FloatAnchor from 'react-float-anchor';
 import type {Options as FloatAnchorOptions} from 'react-float-anchor';
+export type {Options as FloatAnchorOptions} from 'react-float-anchor';
 import MenuItem from './MenuItem';
 
 import type {Direction, Rect} from './types';
@@ -23,25 +24,25 @@ type State = {
 };
 
 export type Props = {
-  menu?: ReactElement<any>;
+  menu: ReactNode;
   positionOptions: FloatAnchorOptions;
-  menuZIndex?: ?string|number;
+  menuZIndex?: string|number;
 
-  onWillOpen?: ?() => void;
-  onDidOpen?: ?() => void;
-  onWillClose?: ?() => void;
+  onWillOpen?: () => void;
+  onDidOpen?: () => void;
+  onWillClose?: () => void;
 
-  className?: ?string;
-  style?: ?Object;
-  highlightedClassName?: ?string;
-  highlightedStyle?: ?Object;
-  index?: ?number;
+  className?: string;
+  style?: Object;
+  highlightedClassName?: string;
+  highlightedStyle?: Object;
+  index?: number;
 
-  openedClassName?: ?string;
-  openedStyle?: ?Object;
+  openedClassName?: string;
+  openedStyle?: Object;
 
-  onItemChosen?: ?(event: ChosenEvent) => void;
-  onHighlightChange?: ?(highlighted: boolean, details: {byKeyboard: ?boolean, prevCursorLocation: ?Rect, direction: ?Direction}) => void;
+  onItemChosen?: (event: ChosenEvent) => void;
+  onHighlightChange?: (highlighted: boolean, details: {byKeyboard?: boolean, prevCursorLocation?: Rect, direction?: Direction}) => void;
 
   children?: ReactNode;
 };
@@ -79,10 +80,10 @@ export default class SubMenuItem extends React.Component<Props, State> {
     opened: false
   };
 
-  _menuItem: MenuItem;
-  _menuInspector: MenuListInspector;
-  _menuContainer: HTMLElement;
-  _floatAnchor: FloatAnchor;
+  _menuItemRef = React.createRef<Class<MenuItem>>();
+  _floatAnchorRef = React.createRef<Class<FloatAnchor>>();
+  _menuInspectorRef = React.createRef<Class<MenuListInspector>>();
+  _menuContainerRef = React.createRef<'div'>();
   _resetMouseLeaveWatcher: Bus<null> = kefirBus();
   _stopper = kefirStopper();
 
@@ -90,23 +91,31 @@ export default class SubMenuItem extends React.Component<Props, State> {
     this._stopper.destroy();
   }
 
-  open(callback?: () => any) {
-    this._menuItem.lockHighlight();
-    if (this.state.opened) return;
+  open(): Promise<void> {
+    const menuItem = this._menuItemRef.current;
+    if (!menuItem) throw new Error();
+
+    menuItem.lockHighlight();
+    if (this.state.opened) return Promise.resolve();
     if (this.props.onWillOpen) this.props.onWillOpen();
-    this.setState({opened: true}, () => {
-      if (this.props.onDidOpen) this.props.onDidOpen();
-      if (callback) callback();
+    menuItem.takeKeyboard();
+    return new Promise(resolve => {
+      this.setState({opened: true}, () => {
+        if (this.props.onDidOpen) this.props.onDidOpen();
+        resolve();
+      });
     });
-    this._menuItem.takeKeyboard();
   }
 
   close() {
+    const menuItem = this._menuItemRef.current;
+    if (!menuItem) throw new Error();
+
     if (!this.state.opened) return;
     if (this.props.onWillClose) this.props.onWillClose();
     this.setState({opened: false});
-    this._menuItem.releaseKeyboard();
-    this._menuItem.unlockHighlight();
+    menuItem.releaseKeyboard();
+    menuItem.unlockHighlight();
   }
 
   toggle() {
@@ -118,23 +127,38 @@ export default class SubMenuItem extends React.Component<Props, State> {
   }
 
   reposition() {
-    this._floatAnchor.reposition();
+    const floatAnchor = this._floatAnchorRef.current;
+    if (!floatAnchor) throw new Error();
+
+    floatAnchor.reposition();
   }
 
   hasHighlight(): boolean {
-    return this._menuItem.hasHighlight();
+    const menuItem = this._menuItemRef.current;
+    if (!menuItem) throw new Error();
+
+    return menuItem.hasHighlight();
   }
 
   highlight(byKeyboard: boolean=true) {
-    this._menuItem.highlight(byKeyboard);
+    const menuItem = this._menuItemRef.current;
+    if (!menuItem) throw new Error();
+
+    menuItem.highlight(byKeyboard);
   }
 
   unhighlight() {
-    this._menuItem.unhighlight();
+    const menuItem = this._menuItemRef.current;
+    if (!menuItem) throw new Error();
+
+    menuItem.unhighlight();
   }
 
   moveCursor(direction: Direction, prevCursorLocation: ?Rect) {
-    this._menuItem.moveCursor(direction, prevCursorLocation);
+    const menuItem = this._menuItemRef.current;
+    if (!menuItem) throw new Error();
+
+    menuItem.moveCursor(direction, prevCursorLocation);
   }
 
   _onHighlightChange(highlighted: boolean, event: Object) {
@@ -155,14 +179,18 @@ export default class SubMenuItem extends React.Component<Props, State> {
   }
 
   _onMouseLeaveItem(event: Object) {
+    const menuItem = this._menuItemRef.current;
+    const menuContainer = this._menuContainerRef.current;
+    if (!menuItem || !menuContainer) throw new Error();
+
     if (!this.state.opened) {
-      this._menuItem.unhighlight();
+      menuItem.unhighlight();
       return;
     }
 
     // If the mouse isn't going toward the menu, then unhighlight ourself.
 
-    const menuRect = this._menuContainer.getBoundingClientRect();
+    const menuRect = menuContainer.getBoundingClientRect();
 
     const startTime = Date.now();
     const startX = event.pageX, startY = event.pageY;
@@ -208,13 +236,16 @@ export default class SubMenuItem extends React.Component<Props, State> {
       .takeUntilBy(this._stopper)
       .onValue(() => {
         this.close();
-        this._menuItem.unhighlight();
+        menuItem.unhighlight();
       });
   }
 
   _mouseEnterMenu() {
+    const menuItem = this._menuItemRef.current;
+    if (!menuItem) throw new Error();
+
     this._resetMouseLeaveWatcher.emit(null);
-    this._menuItem.unlockHighlight();
+    menuItem.unlockHighlight();
   }
 
   render() {
@@ -237,16 +268,13 @@ export default class SubMenuItem extends React.Component<Props, State> {
 
     return (
       <FloatAnchor
-        ref={el => {
-          if (el) this._floatAnchor = el;
-        }}
+        ref={this._floatAnchorRef}
         options={positionOptions}
         zIndex={menuZIndex}
-        anchor={
+        anchor={anchorRef =>
           <MenuItem
-            ref={el => {
-              if (el) this._menuItem = el;
-            }}
+            ref={this._menuItemRef}
+            domRef={anchorRef}
             index={index}
             style={style}
             className={className}
@@ -259,7 +287,9 @@ export default class SubMenuItem extends React.Component<Props, State> {
                 e.stopPropagation();
                 e.preventDefault();
                 this.open();
-                this._menuInspector.moveCursor('down');
+                const menuInspector = this._menuInspectorRef.current;
+                if (!menuInspector) throw new Error();
+                menuInspector.moveCursor('down');
               }
             }}
             onItemChosen={(e: ChosenEvent) => {
@@ -267,7 +297,9 @@ export default class SubMenuItem extends React.Component<Props, State> {
               e.preventDefault();
               this.open();
               if (e.byKeyboard) {
-                this._menuInspector.moveCursor('down');
+                const menuInspector = this._menuInspectorRef.current;
+                if (!menuInspector) throw new Error();
+                menuInspector.moveCursor('down');
               }
             }}
             aria-haspopup={true}
@@ -279,9 +311,7 @@ export default class SubMenuItem extends React.Component<Props, State> {
         float={
           !opened ? null :
             <MenuListInspector
-              ref={el => {
-                if (el) this._menuInspector = el;
-              }}
+              ref={this._menuInspectorRef}
               onLeftPushed={e => {
                 e.stopPropagation();
                 e.preventDefault();
@@ -289,9 +319,7 @@ export default class SubMenuItem extends React.Component<Props, State> {
               }}
             >
               <div
-                ref={el => {
-                  if (el) this._menuContainer = el;
-                }}
+                ref={this._menuContainerRef}
                 onMouseEnter={()=>this._mouseEnterMenu()}
               >
                 {menu}